--- funny/src/main/java/nx/funny/consumer/ProxyFactory.java
@@ -1,12 +1,0 @@
-package nx.funny.consumer;
-
-import nx.funny.registry.ServiceRegistry;
-
-/**
- * 获取待调用服务的代理对象，用于发送调用请求
- */
-public interface ProxyFactory {
-    <T> T getProxy(Class<T> clazz);
-
-    ServiceRegistry getServiceRegistry();
-}
