+++ funny/src/main/java/nx/funny/sampleconsumer/ProxyFactory.java
@@ -1,0 +1,12 @@
+package nx.funny.sampleconsumer;
+
+import nx.funny.registry.ServiceRegistry;
+
+/**
+ * 获取待调用服务的代理对象，用于发送调用请求
+ */
+public interface ProxyFactory {
+    <T> T getProxy(Class<T> clazz);
+
+    ServiceRegistry getServiceRegistry();
+}
