--- funny/src/test/java/nx/funny/consumer/RpcTestInterface.java
@@ -1,7 +1,0 @@
-package nx.funny.consumer;
-
-public interface RpcTestInterface {
-    String sayHello();
-
-    Integer sum(Integer a,Integer b);
-}
