+++ funny/src/test/java/nx/funny/sampleconsumer/RpcTestInterface.java
@@ -1,0 +1,7 @@
+package nx.funny.sampleconsumer;
+
+public interface RpcTestInterface {
+    String sayHello();
+
+    Integer sum(Integer a,Integer b);
+}
