+++ funny/src/test/java/nx/funny/sampleconsumer/RpcTestInterfaceImpl.java
@@ -1,0 +1,29 @@
+package nx.funny.sampleconsumer;
+
+import nx.funny.sampleprovider.ServiceProvider;
+
+public @ServiceProvider("RpcTestInterface")
+class RpcTestInterfaceImpl implements RpcTestInterface{
+    private static int instanceCounter = 0;
+    private String name;
+
+    public RpcTestInterfaceImpl() {
+        instanceCounter++;
+        this.name = "default name";
+    }
+
+    public RpcTestInterfaceImpl(String name) {
+        instanceCounter++;
+        this.name = name;
+    }
+
+    @Override
+    public String sayHello() {
+        return "Hello World "+name+" "+instanceCounter;
+    }
+
+    @Override
+    public Integer sum(Integer a, Integer b) {
+        return a + b;
+    }
+}
