+++ sample-provider/src/main/java/nx/funny/sampleprovider/SampleProviderInterface.java
@@ -1,0 +1,5 @@
+package nx.funny.sampleprovider;
+
+public interface SampleProviderInterface {
+    String sayHello(String name);
+}
