+++ sample-provider/src/main/java/nx/funny/sampleprovider/SampleProviderInterfaceImpl.java
@@ -1,0 +1,8 @@
+package nx.funny.sampleprovider;
+
+@ServiceProvider("SampleProviderInterface")
+public class SampleProviderInterfaceImpl implements SampleProviderInterface {
+    public String sayHello(String name) {
+        return "Hello World "+name;
+    }
+}
