+++ sample-registry/.gitignore
