+++ sample-registry/pom.xml
@@ -2,20 +2,15 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>nx.funny</groupId>
-  <artifactId>sample-consumer</artifactId>
+  <artifactId>sample-registry</artifactId>
   <packaging>jar</packaging>
   <version>1.0-SNAPSHOT</version>
-  <name>sample-consumer</name>
+  <name>sample-registry</name>
   <url>http://maven.apache.org</url>
   <dependencies>
     <dependency>
       <groupId>nx.funny</groupId>
       <artifactId>funny</artifactId>
-      <version>1.0-SNAPSHOT</version>
-    </dependency>
-    <dependency>
-      <groupId>nx.funny</groupId>
-      <artifactId>sample-provider</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
   </dependencies>
